Add User Activity and subject pdf, fix schedule

+++ app/Http/Controllers/Backend/ScheduleController.php
@@ -2,20 +2,34 @@
 
 namespace App\Http\Controllers\Backend;
 
+use App\Http\Controllers\Backend\General\GenerateUserSession;
 use App\Http\Controllers\Controller;
+use App\Models\EMS\ActiveSyandSem;
+use App\Models\EMS\SchoolYear;
+use App\Models\EMS\Specialization;
 use App\Models\Schedule;
+use App\Models\Teacher;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 
 class ScheduleController extends Controller
 {
+    public $path;
+    public $sy_id;
+    public function __construct()
+    {
+        $this->path = 'uploads/schedules/';
+        $this->sy_id = ActiveSyandSem::find(1)->active_SY_id;
+    }
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
         $schedules = Schedule::all();
-        return view('SMS.backend.pages.schedule.index', compact('schedules'));
+        $teachers = Teacher::all();
+        $specializations = Specialization::all();
+        return view('SMS.backend.pages.schedule.index', compact('schedules', 'teachers', 'specializations'));
     }
 
     /**
@@ -34,19 +48,22 @@ class ScheduleController extends Controller
         try {
             $request->validate([
                 'name' => 'required',
-                'image' => 'required|image',
-                'description' => 'required',
+                'image' => 'required|image|mimes:jpg,png,jpeg,gif,svg|max:2048',
+                'teacher_id' => 'required',
+                'specialization_id' => 'required',
             ]);
-            $path = 'uploads/schedules/';
             $image = $request->image;
             $extension = $request->image->extension();
             $file_name = uniqid() . '_' . $request->name . '.' . $extension;
-            $image->storeAs($path, $file_name);
+            $image->storeAs($this->path, $file_name);
             Schedule::create([
                 'name' => $request->name,
-                'image_url' => $path.$file_name,
-                'description' => $request->description,
+                'image_url' => $this->path . $file_name,
+                'teacher_id' => $request->teacher_id,
+                'specialization_id' => $request->specialization_id,
+                'sy_id' => $this->sy_id,
             ]);
+            GenerateUserSession::GenerateSession('Schedule Management', 'Added a Schedule ' . $request->name, auth()->user());
             return redirect()->back()->with('successToast', 'Schedule created successfully');
         } catch (\Throwable $th) {
             return redirect()->back()->with('errorAlert', $th->getMessage());
@@ -75,26 +92,24 @@ class ScheduleController extends Controller
     public function update(Request $request, string $id)
     {
         try {
-            $request->validate([
-                'name' => 'required',
-                'image' => 'required|image|mimes:jpg,png,jpeg,gif,svg|max:2048',
-                'description' => 'required',
-            ]);
-            $path = 'uploads/schedules/';
-            $image = $request->image;
-            $extension = $request->image->extension();
-            $file_name = uniqid() . '_' . $request->name . '.' . $extension;
-            $image->storeAs($path, $file_name);
             $schedule = Schedule::findOrFail($id);
-            if (Storage::exists($schedule->image_url)) {
-                // delete image
-                Storage::delete($schedule->image_url);
+            if ($request->hasFile('image')) {
+                $image = $request->image;
+                $extension = $request->image->extension();
+                $file_name = uniqid() . '_' . $request->name . '.' . $extension;
+                $image->storeAs($this->path, $file_name);
+                if (Storage::exists($schedule->image_url)) {
+                    // delete image
+                    Storage::delete($schedule->image_url);
+                }
             }
             $schedule->update([
                 'name' => $request->name,
-                'image_url' => $path.$file_name,
-                'description' => $request->description,
+                'image_url' => ($request->hasFile('image')) ? $this->path . $file_name : $schedule->image_url,
+                'teacher_id' => $request->teacher_id,
+                'specialization_id' => $request->specialization_id,
             ]);
+            GenerateUserSession::GenerateSession('Schedule Management', 'Updated Schedule ' . $schedule->name, auth()->user());
             return redirect()->back()->with('successToast', 'Schedule updated successfully');
         } catch (\Throwable $th) {
             return redirect()->back()->with('errorAlert', $th->getMessage());
@@ -108,7 +123,13 @@ class ScheduleController extends Controller
     {
         try {
             $schedule = Schedule::findOrFail($id);
+            // check if image exists
+            if (Storage::exists($schedule->image_url)) {
+                // delete image
+                Storage::delete($schedule->image_url);
+            }
             $schedule->delete();
+            GenerateUserSession::GenerateSession('Schedule Management', 'Deleted Schedule ' . $schedule->name, auth()->user());
             return redirect()->back()->with('successToast', 'Schedule deleted successfully');
         } catch (\Throwable $th) {
             return redirect()->back()->with('errorAlert', $th->getMessage());
